# docs/api/openapi-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "DocReader PoC API",
    "description": "Local proof of concept for document ingestion, OCR, classification and extraction.\n\n**There is no authentication.** Every endpoint is anonymous, the Swagger page is public and\nno bearer token is required. Do not expose this API to the internet and do not send real\npersonal documents to it. Use synthetic, masked or authorized files only.\n\nConventions: JSON in camelCase, timestamps in UTC ISO 8601, errors as\n`application/problem+json`, and an `X-Correlation-Id` header on every response.\n\nStage 3 of the execution plan delivers upload, listing, detail, content, deletion, OCR\n(PP-OCRv5 on CPU, see ADR 0002), the raw text and the structured result of seven document\ntypes: CPF card, CIN/RG, CNH, proof of address, CNPJ card, CCMEI and social contract.\nOther document types are read as raw text only.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/retention-policies": {
      "post": {
        "tags": [
          "RetentionPolicies"
        ],
        "summary": "Creates a retention policy for a document type, a product or service, or both.",
        "description": "Precedence, most specific first: type and product together, then the product alone, then the type alone,\nthen the global policy (both omitted), which always exists. There is one policy per scope.\nCreating or changing a policy does not recalculate the purge date of documents that already have one:\nthey pick it up when reprocessed.",
        "requestBody": {
          "description": "The scope and the retention in days.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRetentionPolicyRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRetentionPolicyRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRetentionPolicyRequest"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unexpected-error",
                  "title": "Unexpected error",
                  "status": 500,
                  "detail": "The request could not be completed. Check the API logs using the correlation id of this response.",
                  "instance": "/api/v1/retention-policies",
                  "errorCode": "UNEXPECTED_ERROR",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "201": {
            "description": "Created. The `Location` header points at it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionPolicyResponse"
                }
              }
            }
          },
          "400": {
            "description": "The document type is unknown or the days are out of range.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/validation-failed",
                  "title": "One or more validation errors occurred",
                  "status": 400,
                  "detail": "Check the errors property for the fields that were refused.",
                  "instance": "/api/v1/retention-policies",
                  "errorCode": "VALIDATION_FAILED",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "409": {
            "description": "A policy for this scope already exists (there is exactly one global policy).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/conflict",
                  "title": "Conflict",
                  "status": 409,
                  "detail": "A retention policy for this document type and product or service already exists. Change it instead of creating another.",
                  "instance": "/api/v1/retention-policies",
                  "errorCode": "RETENTION_POLICY_EXISTS",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "422": {
            "description": "The product or service does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unprocessable-content",
                  "title": "Content cannot be processed",
                  "status": 422,
                  "detail": "There is no product or service with the id 0199c1f0-1111-7a10-9c44-2f1d8e6b4a21.",
                  "instance": "/api/v1/retention-policies",
                  "errorCode": "PRODUCT_SERVICE_NOT_FOUND",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "RetentionPolicies"
        ],
        "summary": "Lists the retention policies, the global one first.",
        "parameters": [
          {
            "name": "documentType",
            "in": "query",
            "schema": {
              "maxLength": 64,
              "type": "string"
            }
          },
          {
            "name": "productServiceId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "maximum": 500,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unexpected-error",
                  "title": "Unexpected error",
                  "status": 500,
                  "detail": "The request could not be completed. Check the API logs using the correlation id of this response.",
                  "instance": "/api/v1/retention-policies",
                  "errorCode": "UNEXPECTED_ERROR",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "200": {
            "description": "One page of policies.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionPolicyResponsePagedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/retention-policies/{id}": {
      "get": {
        "tags": [
          "RetentionPolicies"
        ],
        "summary": "Returns one retention policy.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identity.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unexpected-error",
                  "title": "Unexpected error",
                  "status": 500,
                  "detail": "The request could not be completed. Check the API logs using the correlation id of this response.",
                  "instance": "/api/v1/retention-policies/{id}",
                  "errorCode": "UNEXPECTED_ERROR",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "200": {
            "description": "The policy.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionPolicyResponse"
                }
              }
            }
          },
          "404": {
            "description": "There is none with this id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/document-not-found",
                  "title": "Resource not found",
                  "status": 404,
                  "detail": "No retention policy matches the requested identifier.",
                  "instance": "/api/v1/retention-policies/{id}",
                  "errorCode": "RETENTION_POLICY_NOT_FOUND",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RetentionPolicies"
        ],
        "summary": "Changes the number of days of a policy.",
        "description": "The scope is the identity of a policy and cannot change. Documents that already have a purge date keep it until they are reprocessed.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identity.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The new retention in days.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRetentionPolicyRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRetentionPolicyRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRetentionPolicyRequest"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unexpected-error",
                  "title": "Unexpected error",
                  "status": 500,
                  "detail": "The request could not be completed. Check the API logs using the correlation id of this response.",
                  "instance": "/api/v1/retention-policies/{id}",
                  "errorCode": "UNEXPECTED_ERROR",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "200": {
            "description": "Updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionPolicyResponse"
                }
              }
            }
          },
          "400": {
            "description": "The days are out of range.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/validation-failed",
                  "title": "One or more validation errors occurred",
                  "status": 400,
                  "detail": "Check the errors property for the fields that were refused.",
                  "instance": "/api/v1/retention-policies/{id}",
                  "errorCode": "VALIDATION_FAILED",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "404": {
            "description": "There is none with this id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/document-not-found",
                  "title": "Resource not found",
                  "status": 404,
                  "detail": "No retention policy matches the requested identifier.",
                  "instance": "/api/v1/retention-policies/{id}",
                  "errorCode": "RETENTION_POLICY_NOT_FOUND",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "RetentionPolicies"
        ],
        "summary": "DESTRUCTIVE AND IRREVERSIBLE — Deletes a retention policy.",
        "description": "The global policy cannot be deleted. Documents that carried the deleted policy keep their purge date.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identity.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unexpected-error",
                  "title": "Unexpected error",
                  "status": 500,
                  "detail": "The request could not be completed. Check the API logs using the correlation id of this response.",
                  "instance": "/api/v1/retention-policies/{id}",
                  "errorCode": "UNEXPECTED_ERROR",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "204": {
            "description": "Deleted."
          },
          "404": {
            "description": "There is none with this id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/document-not-found",
                  "title": "Resource not found",
                  "status": 404,
                  "detail": "No retention policy matches the requested identifier.",
                  "instance": "/api/v1/retention-policies/{id}",
                  "errorCode": "RETENTION_POLICY_NOT_FOUND",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "409": {
            "description": "It is the global policy.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/conflict",
                  "title": "Conflict",
                  "status": 409,
                  "detail": "The global retention policy cannot be deleted: it is what applies to a document no other policy covers. Change its retentionDays instead.",
                  "instance": "/api/v1/retention-policies/{id}",
                  "errorCode": "GLOBAL_RETENTION_POLICY_PROTECTED",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRetentionPolicyRequest": {
        "required": [
          "retentionDays"
        ],
        "type": "object",
        "properties": {
          "documentType": {
            "maxLength": 64,
            "type": "string",
            "description": "Document type the policy covers (BR_CPF_CARD, BR_CIN, BR_CNH, BR_PROOF_OF_ADDRESS, BR_CNPJ_CARD, BR_CCMEI or BR_SOCIAL_CONTRACT); omit for any type.",
            "nullable": true
          },
          "productServiceId": {
            "type": "string",
            "description": "Id of the product or service the policy covers; omit for any.",
            "format": "uuid",
            "nullable": true
          },
          "retentionDays": {
            "maximum": 36500,
            "minimum": 1,
            "type": "integer",
            "description": "Days a document is kept, from 1 to 36500.",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ProductServiceReferenceResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identity.",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "description": "Unique code."
          },
          "name": {
            "type": "string",
            "description": "Display name."
          }
        },
        "additionalProperties": false,
        "description": "The product or service a document is linked to."
      },
...
      "RetentionPolicyResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identity.",
            "format": "uuid"
          },
          "documentType": {
            "type": "string",
            "description": "Document type the policy covers, or null for any type.",
            "nullable": true
          },
          "productService": {
            "$ref": "#/components/schemas/ProductServiceReferenceResponse"
          },
          "retentionDays": {
            "type": "integer",
            "description": "Days a document is kept, counted from its upload (or from its last reprocessing).",
            "format": "int32"
          },
          "scope": {
            "$ref": "#/components/schemas/RetentionScope"
          },
          "isGlobal": {
            "type": "boolean",
            "description": "The global policy always exists, covers whatever no other policy covers, and cannot be deleted."
          },
          "createdAt": {
            "type": "string",
            "description": "Creation instant, in UTC.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Last change, in UTC.",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "How long documents are kept before they are purged, for one scope.",
        "example": {
          "id": "0199c1f0-2222-7a10-9c44-2f1d8e6b4a21",
          "documentType": "BR_CNPJ_CARD",
          "productService": {
            "id": "0199c1f0-1111-7a10-9c44-2f1d8e6b4a21",
            "code": "CONTA-PJ",
            "name": "Abertura de conta PJ"
          },
          "retentionDays": 365,
          "scope": "DOCUMENT_TYPE_AND_PRODUCT_SERVICE",
          "isGlobal": false,
          "createdAt": "2026-09-24T22:00:00Z",
          "updatedAt": "2026-09-24T22:00:00Z"
        }
      },
      "RetentionPolicyResponsePagedResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RetentionPolicyResponse"
            },
            "description": "Items of the current page, newest first."
          },
          "page": {
            "type": "integer",
            "description": "One based page number.",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "Number of items per page.",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "description": "Total number of items matching the filters.",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "description": "Number of pages for the current page size.",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Envelope of every paginated listing.",
        "example": {
          "items": [
            {
              "id": "0199c1f0-2222-7a10-9c44-2f1d8e6b4a21",
              "documentType": "BR_CNPJ_CARD",
              "productService": {
                "id": "0199c1f0-1111-7a10-9c44-2f1d8e6b4a21",
                "code": "CONTA-PJ",
                "name": "Abertura de conta PJ"
              },
              "retentionDays": 365,
              "scope": "DOCUMENT_TYPE_AND_PRODUCT_SERVICE",
              "isGlobal": false,
              "createdAt": "2026-09-24T22:00:00Z",
              "updatedAt": "2026-09-24T22:00:00Z"
            }
          ],
          "page": 1,
          "pageSize": 20,
          "totalCount": 1,
          "totalPages": 1
        }
      },
      "RetentionScope": {
        "enum": [
          "GLOBAL",
          "DOCUMENT_TYPE",
          "PRODUCT_SERVICE",
          "DOCUMENT_TYPE_AND_PRODUCT_SERVICE"
        ],
        "type": "string",
        "description": "What a retention policy applies to, from the most general to the most specific."
      },
...
      "UpdateRetentionPolicyRequest": {
        "required": [
          "retentionDays"
        ],
        "type": "object",
        "properties": {
          "retentionDays": {
            "maximum": 36500,
            "minimum": 1,
            "type": "integer",
            "description": "Days a document is kept, from 1 to 36500. The scope of a policy cannot change.",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}